+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat(eventbus): add support for default event bus resource policy
 - feat(doc): add package dependency section in typedoc
 - chore(module): add config parsing module
+- feat(organizations): add support for chatbot policies
 
 ### Fixed